Abstracting relevant functions to abstract parent class. #507
