+++ collect_app/src/main/java/org/odk/collect/android/activities/GeoPointMapActivity.java
@@ -12,26 +12,32 @@
  * the License.
  */
 
-
-
-
 package org.odk.collect.android.activities;
 
+import android.content.DialogInterface;
+import android.content.Intent;
 import android.location.Location;
+import android.location.LocationListener;
 import android.location.LocationManager;
+import android.os.Bundle;
+import android.provider.Settings;
 import android.support.v4.app.FragmentActivity;
 import android.widget.Button;
 import android.widget.TextView;
 import android.app.AlertDialog;
+import org.odk.collect.android.R;
+import org.odk.collect.android.application.Collect;
 import org.odk.collect.android.spatial.MapHelper;
 import android.view.View;
+import java.text.DecimalFormat;
+import java.util.List;
 
 /**
  *  Abstracts the functionalities present in GeoPointGoogleMapActivity and GeoPointOsmMapActivity.
  *  @author mukund.code@gmail.com (Mukund Ananthu)
  */
 
-public abstract class GeoPointMapActivity extends FragmentActivity {
+public abstract class GeoPointMapActivity extends FragmentActivity implements LocationListener {
 
     protected static final String LOCATION_COUNT = "locationCount";
     protected TextView locationStatus;
@@ -57,5 +63,103 @@ public abstract class GeoPointMapActivity extends FragmentActivity {
     protected Boolean draggable = false;
     protected Boolean intentDraggable = false;
     protected Boolean locationFromIntent = false;
+
+    @Override
+    protected void onPause() {
+        super.onPause();
+        if (locationManager != null) {
+            locationManager.removeUpdates(this);
+        }
+    }
+
+    @Override
+    public void onProviderDisabled(String provider) {
+
+    }
+
+    @Override
+    public void onProviderEnabled(String provider) {
+
+    }
+
+    @Override
+    public void onStatusChanged(String provider, int status, Bundle extras) {
+    }
+
+    @Override
+    protected void onStart() {
+        super.onStart();
+        Collect.getInstance().getActivityLogger().logOnStart(this);
+    }
+
+    @Override
+    protected void onStop() {
+        Collect.getInstance().getActivityLogger().logOnStop(this);
+        super.onStop();
+    }
+
+    protected String truncateFloat(float f) {
+        return new DecimalFormat("#.##").format(f);
+    }
+
+    protected void upMyLocationOverlayLayers() {
+        // make sure we have a good location provider before continuing
+        locationCountNum = 0;
+        List<String> providers = locationManager.getProviders(true);
+        for (String provider : providers) {
+            if (provider.equalsIgnoreCase(LocationManager.GPS_PROVIDER)) {
+                gpsOn = true;
+                locationManager.requestLocationUpdates(LocationManager.GPS_PROVIDER, 0, 0, this);
+                locationCountFoundLimit = 0;
+            } else if (provider.equalsIgnoreCase(LocationManager.NETWORK_PROVIDER)) {
+                // Only if GPS Provider is not available use network location. bug (well know
+                // android bug) http://stackoverflow
+                // .com/questions/6719207/locationmanager-returns-old-cached-wifi-location-with
+                // -current-timestamp
+                networkOn = true;
+                locationCountFoundLimit =
+                        1; // increase count due to network location bug (well know android bug)
+                locationManager.requestLocationUpdates(LocationManager.NETWORK_PROVIDER, 0, 0,
+                        this);
+            }
+        }
+        //showLocationButton.setClickable(marker != null);
+        if (!gpsOn && !networkOn) {
+            showGPSDisabledAlertToUser();
+        } else {
+            overlayMyLocationLayers();
+        }
+    }
+
+    protected abstract void overlayMyLocationLayers();
+
+    protected void showGPSDisabledAlertToUser() {
+        AlertDialog.Builder alertDialogBuilder = new AlertDialog.Builder(this);
+        alertDialogBuilder.setMessage(getString(R.string.gps_enable_message))
+                .setCancelable(false)
+                .setPositiveButton(getString(R.string.enable_gps),
+                        new DialogInterface.OnClickListener() {
+                            public void onClick(DialogInterface dialog, int id) {
+                                startActivityForResult(
+                                        new Intent(Settings.ACTION_LOCATION_SOURCE_SETTINGS), 0);
+                            }
+                        });
+        alertDialogBuilder.setNegativeButton(getString(R.string.cancel),
+                new DialogInterface.OnClickListener() {
+                    public void onClick(DialogInterface dialog, int id) {
+                        dialog.cancel();
+                    }
+                });
+        AlertDialog alert = alertDialogBuilder.create();
+        alert.show();
+    }
+
+    abstract public void onLocationChanged(Location location);
+
+    abstract protected void zoomToLocation();
+
+    abstract protected void zoomToPoint();
+
+    abstract public void showZoomDialog();
 
 }
